moved definition of damage values from DamageValuePicker to UnitEntity
